+++ commerce/delivery/src/main/java/ru/yandex/practicum/commerce/delivery/service/DeliveryService.java
@@ -1,13 +1,15 @@
 package ru.yandex.practicum.commerce.delivery.service;
 
 import ru.yandex.practicum.commerce.dto.delivery.DeliveryDto;
+
+import java.math.BigDecimal;
 import java.util.UUID;
 
 public interface DeliveryService {
 
     DeliveryDto planDelivery(DeliveryDto deliveryDto);
 
-    Double calculateDeliveryCost(DeliveryDto deliveryDto);
+    BigDecimal calculateDeliveryCost(DeliveryDto deliveryDto);
 
     void deliverySuccess(UUID deliveryId);
